+++ Trarizon.Library/Helpers/AsyncHelper.NullableValueTaskAwaiter.cs
@@ -5,19 +5,6 @@ using Trarizon.Library.Wrappers;
 namespace Trarizon.Library.Helpers;
 partial class AsyncHelper
 {
-    public readonly struct NullableValueTaskAwaiter : INotifyCompletion, ICriticalNotifyCompletion
-    {
-        private readonly ValueTaskAwaiter? _valueTaskAwaiter;
-        
-        [FriendAccess(typeof(AsyncHelper))]
-        internal NullableValueTaskAwaiter(ValueTask? valueTask) => _valueTaskAwaiter = valueTask?.GetAwaiter();
-        
-        public bool IsCompleted => _valueTaskAwaiter?.IsCompleted ?? true;
-        public void OnCompleted(Action continuation) => _valueTaskAwaiter?.OnCompleted(continuation);
-        public void GetResult() => _valueTaskAwaiter?.GetResult();
-        public void UnsafeOnCompleted(Action continuation) => _valueTaskAwaiter?.UnsafeOnCompleted(continuation);
-    }
-
     public readonly struct NullableValueTaskAwaiter<T> : INotifyCompletion, ICriticalNotifyCompletion
     {
         private readonly ValueTaskAwaiter<T>? _valueTaskAwaiter;